Fixed Inf, added stub functions, added memory mapping
Fixed the INF to correctly install the driver.
Added a lot of stub functions to the DXGK initialization function.
Changed the WDDM version to the minimum of Vista.
This is with the goal of getting it to load.
Added memory mapping.
Added ResetDevice function.
For some reason this still isn't loading, reportedly because it can't find the entrypoint...