Added logging statements
Added a lot of logging to the driver to figure why it's not loading correctly.
Allowed the logger to manually allocate a large block if it fails with the statically allocated block.
Added dummy QueryInterface function.
Added __security_init_cookie() to DriverEntry.

"Successful" starting of the driver is currently rather unpredictable (or more accurately, it's very predictable, but annoyingly tedious to get it to "start"). The PCI BUS_INTERFACE_STANDARD is not found most of the time. To get it to find I have to uninstall all of my drivers, reboot the computer, and then reinstall the driver, and sometimes reboot again. The key here is that the Microsoft Basic Display Adapter needs to exist, seemingly to respond the IRP to get the PCI BUS_INTERFACE_STANDARD. 
If it successfully loads, several interfaces are queried (but we just don't report any), and then the device fails our adapter check (probably because the vendor ID is wrong), something to fix tomorrow.