Fixed Vendor ID, Added logging of FileDevice Types, Fixed StartDevice
Changed the Vendor ID to FFFD, this now allows the driver to pass the check device function.
Added stringification for DEVICE_OBJECT::DeviceType.
Fixed Start Device querying the name of the device. It now no longer BSOD's when the buffer is too small (something which is an expected event, but apparently returns an error instead of success when the driver is obviously checking the length), also properly handles freeing of the buffer in several places (previously it was freeing with regard for whether it was actually statically allocated).