Worked on debugging video present networks
Added logging to view video present networks.
The format of the data is
__TYPE__:__HANDLE__:__PARENT__:__CONSTRAINTS__;

Types:
V - Video Present Network
S - Source
T - Target
SM - Source Mode
TM - Target Mode
MM - Monitor Mode

This was simpler to code up than JSON, I might create a viewer later.
Fixed logging of messages greater than 512 characters.
Fixed the page handling of GsSetVidPnSourceAddress.
Changed the production of the target and source modes to be more similar to the monitor modes.

As of now, a VidPN is not being committed, but the D3D10 test application is loading the user mode driver!