Expanded D3D10 testing
Created a window and started the swapchain.
Made debug attempt to load the driver. This way the initial log on adapter creation is performed.
When debugging is activated the UMD will configure the Console. We don't do this until debug is active, since it probably fucks with how the user configured their console. We still need to configure it because we statically linked TauUtils in.