Worked on mapping memory
Attempted a test where I allocated a texture in UM and wrote to it directly. Part of the test was limiting the amount of visible VRAM to the exact size of the framebuffer. Unfortunately it did not work, since it was a dynamic texture, technically it is supposed to map a staging buffer, and I can't directly map a texture until D3D11. I could probably do a no-overwrite buffer though.

On the plus side this exposed some bugs in how I was managing segments that I've now fixed.